PP-4678 Move Cypress specs into folders
- Specs were mostly in `users` folder, even though most weren't to do
  with users. Moved into more appropriate folders
- Move cookie-monster.js into plugins folder, where it belongs
