+++ test/cypress/integration/settings/email_notifications_spec.js
