+++ test/cypress/integration/settings/merchant_details_spec.js
