+++ test/cypress/integration/transactions/transaction_details_spec.js
