+++ test/cypress/plugins/cookie-monster.js
